Added Conditional to check if disk should be brought online.
This fixes a bug where if you want to bring a disk offline by using the parameter "online" set to "false/no" it would bring the disk online again anyways. All needed is the added conditional to fix this, where the parameter gets taken in to account at the end.