Update win_hotfix.ps1
Thank you for the suggestion on $cab_source_filename. I have included that as suggested.
We should be good with not creating a install_result object to store both the property results of Add-WindowsPackage as i see the results of the $hotfix_metadata.path are ordered and the first install of windows package SSU would result with restart not required and the next windows package of security updates should result in restart required and this should reflect in the final value for the variable $result.reboot_required. Let me know your opinion on that.