Update win_hotfix.ps1
I have changed the psobject to hashtable in line 118 and this captures the output from the function as needed for the add-windowspackage. Also line 226 i have put the foreach statement within the install result as the SSU installation always results in reboot not required and the final result of installing the hotfix should get stored in the variable $install_result.