win_psrepository_copy - improve changed check
Improves the method used by `win_psrepository_copy` to determine if any
changes are needed by not relying on the hashtable ordering
implementation. This is not strictly defined and newer .NET runtimes
have issues with how this gets serialized. Using a SortedDictionary
gives us a better guarantee of the serialized output and thus better
idempotency check.